pyproject.toml
[tool.setuptools.packages.find]
include = ["ainfinity*"]
[project]
name = "ainfinity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.4.1",
    "deepspeed>=0.18.2",
    "evaluate>=0.4.6",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "ninja>=1.13.0",
    "omegaconf>=2.3.0",
    "packaging>=25.0",
    "peft>=0.18.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "skypilot[aws,gcp,runpod,vast]>=0.10.5",
    "tensorboard>=2.20.0",
    "torch==2.9.1",
    "tqdm>=4.67.1",
    "pre-commit>=4.5.0",
    "transformers>=4.57.1",
    "trl==0.18.2",
    "vastai-sdk>=0.1.12",
    "wandb>=0.23.0",
    "pytest>=9.0.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "types-pyyaml>=6.0.12.20250915",
]

gpu = [
    "bitsandbytes>=0.48.2",
    "liger-kernel>=0.6.4"
]
